+++ src/parse/extra.rs
@@ -1,4 +1,4 @@
-use crate::ast::SrcSpan;
+use crate::{ast::SrcSpan, GleamExpect};
 
 #[derive(Debug, PartialEq)]
 pub struct ModuleExtra {
@@ -29,7 +29,10 @@ impl<'a> From<(&SrcSpan, &'a str)> for Comment<'a> {
     fn from(src: (&SrcSpan, &'a str)) -> Comment<'a> {
         Comment {
             start: src.0.start,
-            content: &src.1[src.0.start..src.0.end],
+            content: src
+                .1
+                .get(src.0.start..src.0.end)
+                .gleam_expect("From span to comment"),
         }
     }
 }