+++ src/type_/fields.rs
@@ -1,5 +1,8 @@
 use super::Error;
-use crate::ast::{CallArg, SrcSpan};
+use crate::{
+    ast::{CallArg, SrcSpan},
+    GleamExpect,
+};
 use itertools::Itertools;
 use std::collections::{HashMap, HashSet};
 
@@ -70,9 +73,19 @@ impl FieldMap {
 
         let mut i = 0;
         while i < args.len() {
-            let (label, &location) = match &args[i].label {
+            let (label, &location) = match &args
+                .get(i)
+                .gleam_expect("Field indexing to get label")
+                .label
+            {
                 // A labelled argument, we may need to reposition it in the array vector
-                Some(l) => (l, &args[i].location),
+                Some(l) => (
+                    l,
+                    &args
+                        .get(i)
+                        .gleam_expect("Indexing in labelled field reordering")
+                        .location,
+                ),
 
                 // Not a labelled argument
                 None => {