+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Gleam can now run eunit without an external build tool.
 - Gleam can now run an Erlang shell without an external build tool.
 - Projects without rebar3 can be generated using the `gleam-lib` template.
+- A compile time error is now raised when multiple module level constants with the same name are defined.
+- Fixed a bug where declaring a type constructor using reserved erlang keyword in its fields results in invalid erlang code being generated.
 
 ## v0.15.1 -2021-05-07